Refactor module organization.
