--- crates/circuit/src/classical/visitors.rs
@@ -1,65 +1,0 @@
-// This code is part of Qiskit.
-//
-// (C) Copyright IBM 2025
-//
-// This code is licensed under the Apache License, Version 2.0. You may
-// obtain a copy of this license in the LICENSE.txt file in the root directory
-// of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-//
-// Any modifications or derivative works of this code must retain this
-// copyright notice, and modified files need to carry a notice indicating
-// that they have been altered from the originals.
-
-use crate::classical::expr::{Expr, Var};
-
-struct ExprIterator<'a> {
-    stack: Vec<&'a Expr>,
-}
-
-impl<'a> Iterator for ExprIterator<'a> {
-    type Item = &'a Expr;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let Some(expr) = self.stack.pop() else {
-            return None;
-        };
-
-        match expr {
-            Expr::Unary(u) => {
-                self.stack.push(&u.operand);
-            }
-            Expr::Binary(b) => {
-                self.stack.push(&b.left);
-                self.stack.push(&b.right);
-            }
-            Expr::Cast(c) => self.stack.push(&c.operand),
-            Expr::Value(_) => {}
-            Expr::Var(_) => {}
-            Expr::Stretch(_) => {}
-            Expr::Index(i) => {
-                self.stack.push(&i.index);
-                self.stack.push(&i.target);
-            }
-        }
-        Some(expr)
-    }
-}
-
-struct VarIterator<'a>(ExprIterator<'a>);
-
-impl<'a> Iterator for VarIterator<'a> {
-    type Item = &'a Var;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        while let Some(expr) = self.0.next() {
-            if let Expr::Var(v) = expr {
-                return Some(v);
-            }
-        }
-        None
-    }
-}
-
-pub fn iter_vars(expr: &Expr) -> impl Iterator<Item = &Var> {
-    VarIterator(ExprIterator { stack: vec![expr] })
-}
